# src/main/java/org/nana/api/SearchResource.java
package org.nana.api;

import jakarta.inject.Inject;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.MediaType;
import java.util.List;
import org.eclipse.microprofile.openapi.annotations.Operation;
import org.nana.api.ApiDtos.SearchResult;
import org.nana.service.SearchService;

@Path("/api/search")
@Produces(MediaType.APPLICATION_JSON)
public class SearchResource {

    @Inject
    SearchService searchService;

    @GET
    @Operation(operationId = "searchBooks", summary = "Search ebooks on Anna's Archive")
    public List<SearchResult> search(
            @QueryParam("q") String query,
            @QueryParam("lang") String language,
            @QueryParam("ext") String extension,
            @QueryParam("content") String content) {
        if (query == null || query.isBlank()) {
            throw ApiException.badRequest("q is required");
        }
        return searchService.search(query, language, extension, content);
    }
}