# src/main/webui/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "DownloadDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "md5" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "author" : {
            "type" : "string"
          },
          "extension" : {
            "type" : "string"
          },
          "requestedBy" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/DownloadStatus"
          },
          "requestedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "startedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "finishedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "filePath" : {
            "type" : "string"
          },
          "sizeBytes" : {
            "type" : "integer",
            "format" : "int64"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "domainIndexUsed" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "DownloadPage" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DownloadDto"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "DownloadRequest" : {
        "type" : "object",
        "required" : [ "md5", "title" ],
        "properties" : {
          "md5" : {
            "type" : "string",
            "pattern" : "[0-9a-fA-F]{32}"
          },
          "title" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "author" : {
            "type" : "string"
          },
          "extension" : {
            "type" : "string"
          }
        }
      },
      "DownloadStatus" : {
        "type" : "string",
        "enum" : [ "PENDING", "DOWNLOADING", "SUCCESS", "FAILED" ]
      },
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
      "SearchResult" : {
        "type" : "object",
        "properties" : {
          "md5" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "author" : {
            "type" : "string"
          },
          "extension" : {
            "type" : "string"
          },
          "sizeBytes" : {
            "type" : "integer",
            "format" : "int64"
          },
          "language" : {
            "type" : "string"
          },
          "year" : {
            "type" : "integer",
            "format" : "int32"
          },
          "coverUrl" : {
            "type" : "string"
          }
        }
      }
    }
  },
  "paths" : {
    "/api/downloads" : {
      "get" : {
        "summary" : "List downloads, most recent first",
        "operationId" : "listDownloads",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DownloadPage"
                }
              }
            }
          }
        },
        "tags" : [ "Download Resource" ]
      },
      "post" : {
        "summary" : "Queue a server-side ebook download",
        "operationId" : "createDownload",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Download accepted and queued",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DownloadDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Download Resource" ]
      }
    },
    "/api/downloads/{id}" : {
      "get" : {
        "summary" : "Get one download",
        "operationId" : "getDownload",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DownloadDto"
                }
              }
            }
          }
        },
        "tags" : [ "Download Resource" ]
      }
    },
    "/api/search" : {
      "get" : {
        "summary" : "Search ebooks on Anna's Archive",
        "operationId" : "searchBooks",
        "parameters" : [ {
          "name" : "content",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ext",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "lang",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "q",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SearchResult"
                  }
                }
              }
            }
          }
        },
        "tags" : [ "Search Resource" ]
      }
    }
  },
  "info" : {
    "title" : "nana API",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}